moved check page values function to a utilities file so it can be re-used
